small changes to the speed conversion

+++ ROS/dextrobot_ws/src/dextrobot_controller/src/arduino/base_controller/src/Dextrobot.cpp
@@ -178,10 +178,10 @@ void Dextrobot::stop(){
 
 // Converts the velocities provided via geometry_msgs/Twist from m/s 
 // to steps/s that will be provided to the stepper motors
-int Dextrobot::convertToStepsPerSecond(float ms){
+float Dextrobot::convertToStepsPerSecond(float ms){
   // float time = (2*3.1416*WHEEL_RADIUS)/ms; // seconds for one rotation
   // float rpm = 60/time; // rotation per min
   // int stepsPerSecond = (int)(rpm/60)*STEP_PER_REVOLUTION; 
   // compact form
-  return (int)(ms/(2*3.1416*WHEEL_RADIUS))*STEP_PER_REVOLUTION;
+  return (ms/(2*3.1416*WHEEL_RADIUS))*STEP_PER_REVOLUTION;
 }
